+++ OzStripsGUI/Controls/BarCreator.cs
@@ -28,6 +28,7 @@ public partial class BarCreator : UserControl
 
         _bartypes.Add("XXX CROSSING XXX", 3);
         _bartypes.Add("Runway ## Released to SMC", 1);
+        _bartypes.Add("CIRCUIT AREA RELEASED", 1);
         _bartypes.Add("Autorelease Suspended", 2);
         _bartypes.Add("Standby For Ground", 1);
         _bartypes.Add("W - ###.# E - ###.#", 1);