+++ OzStripsGUI/DTO/StripDTO.cs
@@ -8,14 +8,8 @@ namespace MaxRumsey.OzStripsPlugin.Gui.DTO;
 /// <summary>
 /// Represents a controller strip with various operational details.
 /// </summary>
-public class StripControllerDTO
+public class StripDTO
 {
-    /// <summary>
-    /// Gets or sets the aircraft identifier.
-    /// </summary>
-    [JsonPropertyName("acid")]
-    public string acid { get; set; } = string.Empty;
-
     /// <summary>
     /// Gets or sets the bay information for the strip.
     /// </summary>
@@ -69,4 +63,14 @@ public class StripControllerDTO
     /// </summary>
     [JsonPropertyName("ready")]
     public bool ready { get; set; }
+
+    /// <summary>
+    /// Gets or sets the strip key which uniquely identifies the strip.
+    /// </summary>
+    public StripKey StripKey { get; set; }
+
+    /// <summary>
+    /// Gets or sets the strip overriden type.
+    /// </summary>
+    public StripType OverrideStripType { get; set; } = StripType.UNKNOWN;
 }